requirements.txt
asgiref==3.3.1
Django==3.1.5
django-registration-redux==2.9
djangorestframework==3.12.2
psycopg2==2.7.7
pytz==2020.5
sqlparse==0.4.1